fix: component loc
home -> club/challenge
feed -> feed
